Log event items in the treasure chest section
